REST 03/restful01/requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
Django==3.0
django-filter==2.4.0
djangorestframework==3.11.0
httpie==2.2.0
idna==2.10
isort==5.6.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.6
Pygments==2.7.1
pylint==2.6.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
wrapt==1.12.1
